Using sgutils framework, icon_256 (#20)
* Using sgutils framework, icon_256

* Fixed Qt.py for Katana < 3.1 (PyQt4, SIP v1 ONLY)

* Moved out QtPyImporter

* Updated changelogs

* Updated config ymls, editorconfig
