+++ info.yml
@@ -36,3 +36,6 @@ description: "Shotgun Pipeline Toolkit integration in Katana"
 # Required minimum versions for this item to run
 requires_shotgun_version:
 requires_core_version: "v0.12.5"
+
+frameworks:
+    - {"name": "tk-framework-shotgunutils", "version": "v5.x.x", "minimum_version": "v5.1.0"}
